+++ scripts/generate_grammars.py
@@ -10,6 +10,7 @@ from pathlib import Path
 ROOT = Path(__file__).resolve().parents[1]
 SCHEMA = "https://raw.githubusercontent.com/martinring/tmlanguage/master/tmlanguage.json"
 ETA_CLOSE_LOOKAHEAD = r"(?=(\s*)(-|_)?(%>))"
+ETA_DELIMITER_SCOPE = "punctuation.section.embedded.eta"
 
 ETA_REPOSITORY = {
     "eta-js-body": {
@@ -112,17 +113,17 @@ ETA_REPOSITORY = {
         "name": "meta.embedded.block.eta.js.curly",
         "begin": r"\{",
         "beginCaptures": {
-            "0": {"name": "punctuation.section.embedded.begin.eta"},
+            "0": {"name": ETA_DELIMITER_SCOPE},
         },
         "end": rf"(\}})|{ETA_CLOSE_LOOKAHEAD}",
         "endCaptures": {
-            "1": {"name": "punctuation.section.embedded.end.eta"},
+            "1": {"name": ETA_DELIMITER_SCOPE},
         },
         "patterns": [{"include": "#eta-js"}],
     },
     "eta-js-curly-close": {
         "match": r"\}",
-        "name": "punctuation.section.embedded.end.eta",
+        "name": ETA_DELIMITER_SCOPE,
     },
     "eta-js-paren-group": {
         "name": "meta.group.parens.eta.js",
@@ -193,14 +194,14 @@ ETA_REPOSITORY = {
         "name": "meta.embedded.block.eta.output.escaped",
         "begin": "(<%)(-|_)?(\\s*)(=)",
         "beginCaptures": {
-            "1": {"name": "punctuation.section.embedded.begin.eta"},
+            "1": {"name": ETA_DELIMITER_SCOPE},
             "2": {"name": "keyword.operator.whitespace-control.eta"},
             "4": {"name": "keyword.operator.output.escaped.eta"},
         },
         "end": "(\\s*)(-|_)?(%>)",
         "endCaptures": {
             "2": {"name": "keyword.operator.whitespace-control.eta"},
-            "3": {"name": "punctuation.section.embedded.end.eta"},
+            "3": {"name": ETA_DELIMITER_SCOPE},
         },
         "patterns": [{"include": "#eta-js-body"}],
     },
@@ -208,14 +209,14 @@ ETA_REPOSITORY = {
         "name": "meta.embedded.block.eta.output.raw",
         "begin": "(<%)(-|_)?(\\s*)(~)",
         "beginCaptures": {
-            "1": {"name": "punctuation.section.embedded.begin.eta"},
+            "1": {"name": ETA_DELIMITER_SCOPE},
             "2": {"name": "keyword.operator.whitespace-control.eta"},
             "4": {"name": "keyword.operator.output.raw.eta"},
         },
         "end": "(\\s*)(-|_)?(%>)",
         "endCaptures": {
             "2": {"name": "keyword.operator.whitespace-control.eta"},
-            "3": {"name": "punctuation.section.embedded.end.eta"},
+            "3": {"name": ETA_DELIMITER_SCOPE},
         },
         "patterns": [{"include": "#eta-js-body"}],
     },
@@ -223,14 +224,14 @@ ETA_REPOSITORY = {
         "name": "meta.embedded.block.eta.comment",
         "begin": "(<%)(-|_)?(\\s*)(#)",
         "beginCaptures": {
-            "1": {"name": "punctuation.section.embedded.begin.eta"},
+            "1": {"name": ETA_DELIMITER_SCOPE},
             "2": {"name": "keyword.operator.whitespace-control.eta"},
             "4": {"name": "keyword.operator.comment.eta"},
         },
         "end": "(\\s*)(-|_)?(%>)",
         "endCaptures": {
             "2": {"name": "keyword.operator.whitespace-control.eta"},
-            "3": {"name": "punctuation.section.embedded.end.eta"},
+            "3": {"name": ETA_DELIMITER_SCOPE},
         },
         "contentName": "comment.block.eta",
     },
@@ -238,13 +239,13 @@ ETA_REPOSITORY = {
         "name": "meta.embedded.block.eta.code",
         "begin": "(<%)(-|_)?(?!\\s*[=~#])",
         "beginCaptures": {
-            "1": {"name": "punctuation.section.embedded.begin.eta"},
+            "1": {"name": ETA_DELIMITER_SCOPE},
             "2": {"name": "keyword.operator.whitespace-control.eta"},
         },
         "end": "(\\s*)(-|_)?(%>)",
         "endCaptures": {
             "2": {"name": "keyword.operator.whitespace-control.eta"},
-            "3": {"name": "punctuation.section.embedded.end.eta"},
+            "3": {"name": ETA_DELIMITER_SCOPE},
         },
         "patterns": [{"include": "#eta-js-body"}],
     },